refactor(test): split oversized test files to satisfy quench cloc limits
Split three test files that exceeded the 500-line/6000-token quench limits:
- beadsapi/client_test.go (1007 lines) -> client_test.go + client_list_test.go + client_update_test.go
- bridge/sse_test.go (612 lines) -> sse_test.go + sse_edge_test.go
- subscriber/sse_test.go (703 lines) -> sse_test.go + sse_edge_test.go

No test logic changed; shared helpers kept in main test files.
